refactor: refactoring of the middleware.test.js
