+++ test/middleware.test.js
@@ -3,21 +3,16 @@ import path from 'path';
 
 import express from 'express';
 import request from 'supertest';
-import memfs, { createFsFromVolume, Volume } from 'memfs';
 import del from 'del';
 
 import middleware from '../src';
 
 import getCompiler from './helpers/getCompiler';
 import GetLogsPlugin from './helpers/GetLogsPlugin';
-import isWebpack5 from './helpers/isWebpack5';
 
 import webpackConfig from './fixtures/webpack.config';
 import webpackSimpleConfig from './fixtures/webpack.simple.config';
 import webpackMultiConfig from './fixtures/webpack.array.config';
-import webpackWatchOptionsConfig from './fixtures/webpack.watch-options.config';
-import webpackMultiWatchOptionsConfig from './fixtures/webpack.array.watch-options.config';
-import webpackQueryStringConfig from './fixtures/webpack.querystring.config';
 import webpackErrorConfig from './fixtures/webpack.error.config';
 import webpackMultiErrorConfig from './fixtures/webpack.array.error.config';
 import webpackWarningConfig from './fixtures/webpack.warning.config';
@@ -57,573 +52,6 @@ describe('middleware', () => {
       }
     });
   }
-
-  describe('watchOptions option', () => {
-    describe('should work without value', () => {
-      let compiler;
-      let spy;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackConfig);
-
-        spy = jest.spyOn(compiler, 'watch');
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll((done) => {
-        spy.mockRestore();
-
-        close(done);
-      });
-
-      it('should pass arguments to the "watch" method', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            expect(spy).toHaveBeenCalledTimes(1);
-            expect(spy.mock.calls[0][0]).toEqual({});
-
-            return done();
-          });
-      });
-    });
-
-    describe('should respect options from the configuration', () => {
-      let compiler;
-      let spy;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackWatchOptionsConfig);
-
-        spy = jest.spyOn(compiler, 'watch');
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll((done) => {
-        spy.mockRestore();
-
-        close(done);
-      });
-
-      it('should pass arguments to the "watch" method', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(done);
-            }
-
-            expect(spy).toHaveBeenCalledTimes(1);
-            expect(spy.mock.calls[0][0]).toEqual({
-              aggregateTimeout: 300,
-              poll: true,
-            });
-
-            return done(error);
-          });
-      });
-    });
-
-    describe('should respect options from the configuration in multi-compile mode', () => {
-      let compiler;
-      let spy;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackMultiWatchOptionsConfig);
-
-        spy = jest.spyOn(compiler, 'watch');
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll((done) => {
-        spy.mockRestore();
-
-        close(done);
-      });
-
-      it('should pass arguments to the "watch" method', (done) => {
-        request(app)
-          .get('/static-one/bundle.js')
-          .expect(200, (firstError) => {
-            if (firstError) {
-              return done(firstError);
-            }
-
-            return request(app)
-              .get('/static-two/bundle.js')
-              .expect(200, (secondError) => {
-                if (secondError) {
-                  return done(secondError);
-                }
-
-                expect(spy).toHaveBeenCalledTimes(1);
-                expect(spy.mock.calls[0][0]).toEqual([
-                  { aggregateTimeout: 800, poll: false },
-                  { aggregateTimeout: 300, poll: true },
-                ]);
-
-                return done();
-              });
-          });
-      });
-    });
-  });
-
-  describe('writeToDisk option', () => {
-    describe('should work with "true" value', () => {
-      let compiler;
-
-      beforeAll((done) => {
-        compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: path.resolve(__dirname, './outputs/write-to-disk-true'),
-          },
-        });
-
-        instance = middleware(compiler, { writeToDisk: true });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll((done) => {
-        del.sync(path.posix.resolve(__dirname, './outputs/write-to-disk-true'));
-
-        close(done);
-      });
-
-      it('should find the bundle file on disk', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            const bundlePath = path.resolve(
-              __dirname,
-              './outputs/write-to-disk-true/bundle.js'
-            );
-
-            expect(
-              compiler.hooks.assetEmitted.taps.filter(
-                (hook) => hook.name === 'DevMiddleware'
-              ).length
-            ).toBe(1);
-            expect(fs.existsSync(bundlePath)).toBe(true);
-
-            instance.invalidate();
-
-            return compiler.hooks.done.tap(
-              'DevMiddlewareWriteToDiskTest',
-              () => {
-                expect(
-                  compiler.hooks.assetEmitted.taps.filter(
-                    (hook) => hook.name === 'DevMiddleware'
-                  ).length
-                ).toBe(1);
-
-                done();
-              }
-            );
-          });
-      });
-    });
-
-    describe('should work with "false" value', () => {
-      let compiler;
-
-      beforeAll((done) => {
-        compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: path.resolve(__dirname, './outputs/write-to-disk-false'),
-          },
-        });
-
-        instance = middleware(compiler, { writeToDisk: false });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should not find the bundle file on disk', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            const bundlePath = path.resolve(
-              __dirname,
-              './outputs/write-to-disk-false/bundle.js'
-            );
-
-            expect(
-              compiler.hooks.assetEmitted.taps.filter(
-                (hook) => hook.name === 'DevMiddleware'
-              ).length
-            ).toBe(0);
-            expect(fs.existsSync(bundlePath)).toBe(false);
-
-            instance.invalidate();
-
-            return compiler.hooks.done.tap(
-              'DevMiddlewareWriteToDiskTest',
-              () => {
-                expect(
-                  compiler.hooks.assetEmitted.taps.filter(
-                    (hook) => hook.name === 'DevMiddleware'
-                  ).length
-                ).toBe(0);
-
-                done();
-              }
-            );
-          });
-      });
-    });
-
-    describe('should work with "Function" value when it returns "true"', () => {
-      let compiler;
-
-      beforeAll((done) => {
-        compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: path.resolve(
-              __dirname,
-              './outputs/write-to-disk-function-true'
-            ),
-          },
-        });
-
-        instance = middleware(compiler, {
-          writeToDisk: (filePath) => /bundle\.js$/.test(filePath),
-        });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll((done) => {
-        del.sync(
-          path.posix.resolve(__dirname, './outputs/write-to-disk-function-true')
-        );
-
-        close(done);
-      });
-
-      it('should find the bundle file on disk', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            const bundlePath = path.resolve(
-              __dirname,
-              './outputs/write-to-disk-function-true/bundle.js'
-            );
-
-            expect(fs.existsSync(bundlePath)).toBe(true);
-
-            return done();
-          });
-      });
-    });
-
-    describe('should work with "Function" value when it returns "false"', () => {
-      let compiler;
-
-      beforeAll((done) => {
-        compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: path.resolve(
-              __dirname,
-              './outputs/write-to-disk-function-false'
-            ),
-          },
-        });
-
-        instance = middleware(compiler, {
-          writeToDisk: (filePath) => !/bundle\.js$/.test(filePath),
-        });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll((done) => {
-        del.sync(
-          path.posix.resolve(
-            __dirname,
-            './outputs/write-to-disk-function-false'
-          )
-        );
-
-        close(done);
-      });
-
-      it('should not find the bundle file on disk', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            const bundlePath = path.resolve(
-              __dirname,
-              './outputs/write-to-disk-function-false/bundle.js'
-            );
-
-            expect(fs.existsSync(bundlePath)).toBe(false);
-
-            return done();
-          });
-      });
-    });
-
-    describe('should work when assets have query string', () => {
-      let compiler;
-
-      beforeAll((done) => {
-        compiler = getCompiler({
-          ...webpackQueryStringConfig,
-          output: {
-            filename: 'bundle.js?[contenthash]',
-            path: path.resolve(
-              __dirname,
-              './outputs/write-to-disk-query-string'
-            ),
-          },
-        });
-
-        instance = middleware(compiler, { writeToDisk: true });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll((done) => {
-        del.sync(
-          path.posix.resolve(__dirname, './outputs/write-to-disk-query-string')
-        );
-
-        close(done);
-      });
-
-      it('should find the bundle file on disk with no querystring', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            const bundlePath = path.resolve(
-              __dirname,
-              './outputs/write-to-disk-query-string/bundle.js'
-            );
-
-            expect(fs.existsSync(bundlePath)).toBe(true);
-
-            return done();
-          });
-      });
-    });
-
-    describe('should work in multi-compiler mode', () => {
-      let compiler;
-
-      beforeAll((done) => {
-        compiler = getCompiler([
-          {
-            ...webpackMultiWatchOptionsConfig[0],
-            output: {
-              filename: 'bundle.js',
-              path: path.resolve(
-                __dirname,
-                './outputs/write-to-disk-multi-compiler/static-one'
-              ),
-              publicPath: '/static-one/',
-            },
-          },
-          {
-            ...webpackMultiWatchOptionsConfig[1],
-            output: {
-              filename: 'bundle.js',
-              path: path.resolve(
-                __dirname,
-                './outputs/write-to-disk-multi-compiler/static-two'
-              ),
-              publicPath: '/static-two/',
-            },
-          },
-        ]);
-
-        instance = middleware(compiler, { writeToDisk: true });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll((done) => {
-        del.sync(
-          path.posix.resolve(
-            __dirname,
-            './outputs/write-to-disk-multi-compiler/'
-          )
-        );
-
-        close(done);
-      });
-
-      it('should find the bundle files on disk', (done) => {
-        request(app)
-          .get('/static-one/bundle.js')
-          .expect(200, (firstError) => {
-            if (firstError) {
-              return done(firstError);
-            }
-
-            return request(app)
-              .get('/static-two/bundle.js')
-              .expect(200, (secondError) => {
-                if (secondError) {
-                  return done(secondError);
-                }
-                const bundleFiles = [
-                  './outputs/write-to-disk-multi-compiler/static-one/bundle.js',
-                  './outputs/write-to-disk-multi-compiler/static-one/index.html',
-                  './outputs/write-to-disk-multi-compiler/static-one/svg.svg',
-                  './outputs/write-to-disk-multi-compiler/static-two/bundle.js',
-                ];
-
-                for (const bundleFile of bundleFiles) {
-                  const bundlePath = path.resolve(__dirname, bundleFile);
-
-                  expect(fs.existsSync(bundlePath)).toBe(true);
-                }
-
-                return done();
-              });
-          });
-      });
-    });
-
-    describe('should work with "[hash]"/"[fullhash]" in the "output.path" and "output.publicPath" option', () => {
-      let compiler;
-      let hash;
-
-      beforeAll((done) => {
-        compiler = getCompiler({
-          ...webpackConfig,
-          ...{
-            output: {
-              filename: 'bundle.js',
-              publicPath: isWebpack5()
-                ? '/static/[fullhash]/'
-                : '/static/[hash]/',
-              path: isWebpack5()
-                ? path.resolve(
-                    __dirname,
-                    './outputs/write-to-disk-with-hash/dist_[fullhash]'
-                  )
-                : path.resolve(
-                    __dirname,
-                    './outputs/write-to-disk-with-hash/dist_[hash]'
-                  ),
-            },
-          },
-        });
-
-        instance = middleware(compiler, { writeToDisk: true });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(() => {
-          compiler.hooks.afterCompile.tap('wdm-test', ({ hash: h }) => {
-            hash = h;
-            done();
-          });
-        });
-      });
-
-      afterAll((done) => {
-        del.sync(
-          path.posix.resolve(__dirname, './outputs/write-to-disk-with-hash/')
-        );
-
-        close(done);
-      });
-
-      it('should find the bundle file on disk', (done) => {
-        request(app)
-          .get(`/static/${hash}/bundle.js`)
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            const bundlePath = path.resolve(
-              __dirname,
-              `./outputs/write-to-disk-with-hash/dist_${hash}/bundle.js`
-            );
-
-            expect(fs.existsSync(bundlePath)).toBe(true);
-
-            return done();
-          });
-      });
-    });
-  });
 
   describe('methods option', () => {
     let compiler;
@@ -767,171 +195,6 @@ describe('middleware', () => {
 
           return done();
         });
-    });
-  });
-
-  describe('outputFileSystem option', () => {
-    describe('should work with an unspecified value', () => {
-      let compiler;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackConfig);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should use the "memfs" package by default', () => {
-        const { Stats } = memfs;
-
-        expect(new compiler.outputFileSystem.Stats()).toBeInstanceOf(Stats);
-        expect(new instance.context.outputFileSystem.Stats()).toBeInstanceOf(
-          Stats
-        );
-        expect(compiler.outputFileSystem).toHaveProperty('join');
-        expect(compiler.outputFileSystem).toHaveProperty('mkdirp');
-      });
-    });
-
-    describe('should work with the configured value (native fs)', () => {
-      let compiler;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackConfig);
-
-        const configuredFs = fs;
-
-        configuredFs.join = path.join.bind(path);
-        configuredFs.mkdirp = () => {};
-
-        instance = middleware(compiler, {
-          outputFileSystem: configuredFs,
-        });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should use the configurated output filesystem', () => {
-        const { Stats } = fs;
-
-        expect(new compiler.outputFileSystem.Stats()).toBeInstanceOf(Stats);
-        expect(new instance.context.outputFileSystem.Stats()).toBeInstanceOf(
-          Stats
-        );
-        expect(compiler.outputFileSystem).toHaveProperty('join');
-        expect(compiler.outputFileSystem).toHaveProperty('mkdirp');
-      });
-    });
-
-    describe('should work with the configured value (memfs)', () => {
-      let compiler;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackConfig);
-
-        const configuredFs = createFsFromVolume(new Volume());
-
-        configuredFs.join = path.join.bind(path);
-
-        instance = middleware(compiler, {
-          outputFileSystem: configuredFs,
-        });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should use the configured output filesystem', () => {
-        const { Stats } = memfs;
-
-        expect(new compiler.outputFileSystem.Stats()).toBeInstanceOf(Stats);
-        expect(new instance.context.outputFileSystem.Stats()).toBeInstanceOf(
-          Stats
-        );
-        expect(compiler.outputFileSystem).toHaveProperty('join');
-        expect(compiler.outputFileSystem).toHaveProperty('mkdirp');
-      });
-    });
-
-    describe('should work with the configured value in multi-compiler mode (native fs)', () => {
-      let compiler;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackMultiConfig);
-
-        const configuredFs = fs;
-
-        configuredFs.join = path.join.bind(path);
-        configuredFs.mkdirp = () => {};
-
-        instance = middleware(compiler, {
-          outputFileSystem: configuredFs,
-        });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should use configured output filesystems', () => {
-        const { Stats } = fs;
-
-        for (const childCompiler of compiler.compilers) {
-          expect(new childCompiler.outputFileSystem.Stats()).toBeInstanceOf(
-            Stats
-          );
-          expect(childCompiler.outputFileSystem).toHaveProperty('join');
-          expect(childCompiler.outputFileSystem).toHaveProperty('mkdirp');
-        }
-
-        expect(new instance.context.outputFileSystem.Stats()).toBeInstanceOf(
-          Stats
-        );
-        expect(instance.context.outputFileSystem).toHaveProperty('join');
-        expect(instance.context.outputFileSystem).toHaveProperty('mkdirp');
-      });
-    });
-
-    describe('should throw an error on the invalid fs value - no join method', () => {
-      it('should throw an error', () => {
-        expect(() => {
-          const compiler = getCompiler(webpackConfig);
-
-          middleware(compiler, { outputFileSystem: { mkdirp: () => {} } });
-        }).toThrow(
-          'Invalid options: options.outputFileSystem.join() method is expected'
-        );
-      });
-    });
-
-    describe('should throw an error on the invalid fs value - no mkdirp method', () => {
-      it('should throw an error', () => {
-        expect(() => {
-          const compiler = getCompiler(webpackConfig);
-
-          middleware(compiler, { outputFileSystem: { join: () => {} } });
-        }).toThrow(
-          'Invalid options: options.outputFileSystem.mkdirp() method is expected'
-        );
-      });
     });
   });
